Now preserves the histogram axies for all histograms.
This was done in a brute force way. I repeated the same lines multiple times
and looking at it makes me sick to my stomach. That said, it seems to work.

I'm going to spend some time to clean this up and get it looking pretty.

Most important point : The LOGIC is sound.